\name{sonar}
\alias{sonar.r}
\alias{sonar.m}
\docType{data}
\title{
Sonar Returns
}
\description{
The sonar returns were collected from bouncing signals off a metal cylinder and a cylindrically shaped rock, respectively positioned on a sandy ocean floor.\code{sonar.m} contains the sonar returns from the signals that were bounced off the metal cylinder, and \code{sonar.r} contains the sonar returns from the signals bounced off the rock.
}
\usage{data("sonar")}
\format{
\code{sonar.m} contains 111 rows and 60 columns. \code{sonar.r} contains 97 rows and 60 columns.
}

\source{
University of California Irvine Machine Learning Repository
}
\references{
Qiu, Y. and Chen, S. X. (2015) \emph{Journal of the American Statistical Association}. \bold{110.511}, 1160--1174.
}


